requirements.txt
Keras==2.4.3
scipy==1.4.1
tensorflow==2.2.0
opencv_python_headless==4.3.0.36
requests==2.24.0
numpy==1.19.0
fastapi==0.58.1
uvicorn
Pillow==7.2.0
python-multipart

